Postgresql uses smallint instead of tinyint.

+++ lib/db/postgres7.php
@@ -495,7 +495,7 @@ function main_upgrade($oldversion=0) {
     }
     
     if ($oldversion < 2004050400) {     /// add a visible field for events
-        table_column("event", "", "visible", "tinyint", "1", "", "1", "not null", "timeduration");
+        table_column("event", "", "visible", "smallint", "1", "", "1", "not null", "timeduration");
         if ($events = get_records('event')) {
             foreach($events as $event) {
                 if ($moduleid = get_field('modules', 'id', 'name', $event->modulename)) {
